+++ Game/src/World/Terrain.cpp
@@ -18,7 +18,7 @@ terrain::CompoundSurfaceData terrain::CompoundSurfaceData::operator+(const Compo
   return sum;
 }
 
-terrain::CompoundSurfaceData terrain::CompoundSurfaceData::operator*(float x) const
+terrain::CompoundSurfaceData terrain::CompoundSurfaceData::operator*(f32 x) const
 {
   CompoundSurfaceData result = *this;
   result.m_Elevation *= x;
@@ -37,12 +37,12 @@ block::Type terrain::CompoundSurfaceData::getPrimaryBlockType() const
   return m_Components.getPrimary();
 }
 
-std::array<int, 2> terrain::CompoundSurfaceData::getTextureIndices() const
+std::array<i32, 2> terrain::CompoundSurfaceData::getTextureIndices() const
 {
-  std::array<int, 2> textureIndices{};
+  std::array<i32, 2> textureIndices{};
 
-  textureIndices[0] = static_cast<int>(m_Components[0].type.texture(eng::math::Direction::Top));
-  textureIndices[1] = static_cast<int>(m_Components[1].type.texture(eng::math::Direction::Top));
+  textureIndices[0] = static_cast<i32>(m_Components[0].type.texture(eng::math::Direction::Top));
+  textureIndices[1] = static_cast<i32>(m_Components[1].type.texture(eng::math::Direction::Top));
 
   return textureIndices;
 }
@@ -55,12 +55,12 @@ eng::math::Float2 terrain::CompoundSurfaceData::getTextureWeights() const
 
 
 static constexpr length_t c_LargestNoiseScale = 1024 * block::length();
-static constexpr float c_NoiseLacunarity = 2.0f;
+static constexpr f32 c_NoiseLacunarity = 2.0f;
 
-static constexpr int c_MaxCompoundBiomes = 4;
-static constexpr int c_BiomeRegionSize = 8;
-static constexpr int c_RegionRadius = 1;
-static constexpr int c_RegionWidth = 2 * c_RegionRadius + 1;
+static constexpr i32 c_MaxCompoundBiomes = 4;
+static constexpr i32 c_BiomeRegionSize = 8;
+static constexpr i32 c_RegionRadius = 1;
+static constexpr i32 c_RegionWidth = 2 * c_RegionRadius + 1;
 
 using NoiseSamples = BlockArrayRect<noise::OctaveNoiseData<Biome::LocalElevationOctaves()>>;
 using CompoundBiome = CompoundType<Biome::Type, c_MaxCompoundBiomes>;
@@ -72,35 +72,35 @@ struct SurfaceData
   BiomeData biomeData = BlockArrayRect<CompoundBiome>(Chunk::Bounds2D(), eng::AllocationPolicy::ForOverwrite);
 };
 
-static constexpr int c_CacheSize = (2 * c_UnloadDistance + 5) * (2 * c_UnloadDistance + 5);
+static constexpr i32 c_CacheSize = (2 * c_UnloadDistance + 5) * (2 * c_UnloadDistance + 5);
 static eng::threads::LRUCache<GlobalIndex2D, SurfaceData> s_SurfaceDataCache(c_CacheSize);
 static std::mutex s_Mutex;
 
 // From https://stackoverflow.com/questions/664014/what-integer-hash-function-are-good-that-accepts-an-integer-hash-key
-static uint32_t hash(uint32_t n)
+static u32 hash(u32 n)
 {
   n = ((n >> 16) ^ n) * 0x45d9f3b;
   n = ((n >> 16) ^ n) * 0x45d9f3b;
   return (n >> 16) ^ n;
 }
 
-// Returns a random float in the range [0.0, 1.0] based determinisitically on the input n
-static float random(uint32_t n)
+// Returns a random f32 in the range [0.0, 1.0] based determinisitically on the input n
+static f32 random(u32 n)
 {
-  return static_cast<float>(hash(n)) / std::numeric_limits<uint32_t>::max();
+  return static_cast<f32>(hash(n)) / std::numeric_limits<u32>::max();
 }
 
 // Returns a biome type based determinisitically on the input n
-static Biome::Type randomBiome(uint32_t n)
+static Biome::Type randomBiome(u32 n)
 {
   return static_cast<Biome::Type>(hash(n) % Biome::Count());
 }
 
 static std::pair<Biome::Type, eng::math::Float2> getRegionVoronoiPoint(const GlobalIndex2D& regionIndex)
 {
-  uint32_t key = std::hash<GlobalIndex2D>{}(regionIndex);
-  float r = 0.5f * random(key);
-  float theta = 2 * std::numbers::pi_v<float> * random(hash(key));
+  u32 key = static_cast<u32>(std::hash<GlobalIndex2D>{}(regionIndex));
+  f32 r = 0.5f * random(key);
+  f32 theta = 2 * std::numbers::pi_v<f32> * random(hash(key));
 
   Biome::Type biomeType = randomBiome(key);
   eng::math::Float2 relativeLocation(r * std::cos(theta), r * std::sin(theta));
@@ -123,17 +123,17 @@ static CompoundBiome getBiomeData(const eng::math::Vec2& surfaceLocation)
 
       auto [biomeType, voronoiPointPerturbation] = getRegionVoronoiPoint(regionIndex);
       eng::math::Float2 voronoiPointRelativeToQueryRegion = regionCenterRelativeToQueryRegion + voronoiPointPerturbation;
-      float distance = glm::distance(queryLocationRelativeToQueryRegion, voronoiPointRelativeToQueryRegion);
-      float biomeWeight = std::expf(-32 * distance * distance);
+      f32 distance = glm::distance(queryLocationRelativeToQueryRegion, voronoiPointRelativeToQueryRegion);
+      f32 biomeWeight = std::expf(-32 * distance * distance);
 
-      int index = c_RegionWidth * (i + c_RegionRadius) + j + c_RegionRadius;
+      i32 index = c_RegionWidth * (i + c_RegionRadius) + j + c_RegionRadius;
       nearbyBiomes[index] = { biomeType, biomeWeight };
     }
 
   // Combine components of same type
   eng::algo::sort(nearbyBiomes, [](const WeightedBiome& biome) { return eng::toUnderlying(biome.type); }, eng::SortPolicy::Ascending);
-  int lastUniqueElementIndex = 0;
-  for (int i = 1; i < nearbyBiomes.size(); ++i)
+  i32 lastUniqueElementIndex = 0;
+  for (i32 i = 1; i < nearbyBiomes.size(); ++i)
   {
     if (nearbyBiomes[i].type == nearbyBiomes[i - 1].type)
     {
@@ -182,7 +182,7 @@ static void heightMapStage(BlockArrayRect<length_t>& heightMap, const GlobalInde
     for (blockIndex_t j = 0; j < Chunk::Size(); ++j)
     {
       length_t elevation = 0.0;
-      for (int n = 0; n < c_MaxCompoundBiomes; ++n)
+      for (i32 n = 0; n < c_MaxCompoundBiomes; ++n)
       {
         const auto& [biomeType, biomeWeight] = biomeMap[i][j][n];
         if (biomeType == Biome::Type::Null)
@@ -214,16 +214,16 @@ static void foliageStage(BlockArrayBox<block::Type>& composition, const BlockArr
 {
   const auto createTree = [](BlockArrayBox<block::Type>& composition, const BlockIndex& treeIndex, block::Type leafType)
   {
-    int i = treeIndex.i;
-    int j = treeIndex.j;
-    int k = treeIndex.k;
+    i32 i = treeIndex.i;
+    i32 j = treeIndex.j;
+    i32 k = treeIndex.k;
 
-    for (int n = 0; n < 5; ++n)
+    for (i32 n = 0; n < 5; ++n)
       composition[i][j][k + n] = block::ID::OakLog;
 
-    for (int I = -3; I < 3; ++I)
-      for (int J = -3; J < 3; ++J)
-        for (int K = 0; K < 3; ++K)
+    for (i32 I = -3; I < 3; ++I)
+      for (i32 J = -3; J < 3; ++J)
+        for (i32 K = 0; K < 3; ++K)
           if (I * I + J * J + K * K < 9)
             composition[i + I][j + J][k + K + 5] = leafType;
   };
@@ -245,7 +245,7 @@ static void foliageStage(BlockArrayBox<block::Type>& composition, const BlockArr
         length_t heightInChunk = elevation - chunkFloor;
         if (0 < heightInChunk && heightInChunk < Chunk::Length() - 8 * block::length() && 3 < i && i < Chunk::Size() - 4 && 3 < j && j < Chunk::Size() - 4)
         {
-          int random = rand();
+          i32 random = rand();
           if (random % 101 == 0)
           {
             block::ID leafType = random % 2 == 0 ? block::ID::OakLeaves : block::ID::FallLeaves;
